adding some more css to store, and changing product price to take float instead of int. Fixing cancel button on create-account

+++ src/main/java/edu/depaul/cdm/se452/group3/project/entities/Product.java
@@ -20,7 +20,7 @@ public class Product {
     private String productName;
     private String productDescription;
     private String CategoryId;
-    private int Price;
+    private float Price;
     private String imageURL;
 
 }
